pyproject.toml
[tool.poetry]
name = "record-linkage"
version = "0.1.0"
description = ""
authors = ["Rashid Ayoupov <valentaine1234@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
alembic = "^1.12.1"
asyncpg = "^0.29.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
gunicorn = "^21.2.0"
numpy = "^1.26.2"
pandas = "^2.1.3"
scipy = "^1.11.4"
sentence-transformers = "^2.2.2"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
autoflake = "^2.2.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
